+++ packages/firestore/src/local/indexeddb_schema.ts
@@ -29,8 +29,16 @@ import { SnapshotVersion } from '../core/snapshot_version';
  * Schema Version for the Web client:
  * 1. Initial version including Mutation Queue, Query Cache, and Remote Document
  *    Cache
+<<<<<<< HEAD
  * 2. Added targetCount to targetGlobal row.
  * 3. Multi-Tab Support.
+=======
+ * 2. Used to ensure a targetGlobal object exists and add targetCount to it. No
+ *    longer required because migration 3 unconditionally clears it.
+ * 3. Dropped and re-created Query Cache to deal with cache corruption related
+ *    to limbo resolution. Addresses
+ *    https://github.com/firebase/firebase-ios-sdk/issues/1548
+>>>>>>> master
  */
 export const SCHEMA_VERSION = 3;
 
@@ -48,7 +56,11 @@ export function createOrUpgradeDb(
   toVersion: number
 ): PersistencePromise<void> {
   assert(
+<<<<<<< HEAD
     fromVersion < toVersion && fromVersion >= 0 && toVersion <= 3,
+=======
+    fromVersion < toVersion && fromVersion >= 0 && toVersion <= SCHEMA_VERSION,
+>>>>>>> master
     'Unexpected schema upgrade from v${fromVersion} to v{toVersion}.'
   );
   let p = PersistencePromise.resolve();
@@ -60,6 +72,7 @@ export function createOrUpgradeDb(
     createRemoteDocumentCache(db);
   }
 
+<<<<<<< HEAD
   if (fromVersion < 2 && toVersion >= 2) {
     p = p
       .next(() => ensureTargetGlobalExists(txn))
@@ -81,7 +94,22 @@ export function createOrUpgradeDb(
       createClientMetadataStore(db);
       createRemoteDocumentChangesStore(db);
     });
+=======
+  // Migration 2 to populate the targetGlobal object no longer needed since
+  // migration 3 unconditionally clears it.
+
+  let p = PersistencePromise.resolve();
+  if (fromVersion < 3 && toVersion >= 3) {
+    // Brand new clients don't need to drop and recreate--only clients that
+    // potentially have corrupt data.
+    if (fromVersion !== 0) {
+      dropQueryCache(db);
+      createQueryCache(db);
+    }
+    p = p.next(() => writeEmptyTargetGlobalEntry(txn));
+>>>>>>> master
   }
+
   return p;
 }
 
@@ -575,49 +603,30 @@ function createQueryCache(db: IDBDatabase): void {
   db.createObjectStore(DbTargetGlobal.store);
 }
 
+function dropQueryCache(db: IDBDatabase): void {
+  db.deleteObjectStore(DbTargetDocument.store);
+  db.deleteObjectStore(DbTarget.store);
+  db.deleteObjectStore(DbTargetGlobal.store);
+}
+
 /**
- * Counts the number of targets persisted and adds that value to the target
- * global singleton.
+ * Creates the target global singleton row.
+ *
+ * @param {IDBTransaction} txn The version upgrade transaction for indexeddb
  */
-function saveTargetCount(
-  txn: SimpleDbTransaction,
-  metadata: DbTargetGlobal
+function writeEmptyTargetGlobalEntry(
+  txn: SimpleDbTransaction
 ): PersistencePromise<void> {
   const globalStore = txn.store<DbTargetGlobalKey, DbTargetGlobal>(
     DbTargetGlobal.store
   );
-  const targetStore = txn.store<DbTargetKey, DbTarget>(DbTarget.store);
-  return targetStore.count().next(count => {
-    metadata.targetCount = count;
-    return globalStore.put(DbTargetGlobal.key, metadata);
-  });
-}
-
-/**
- * Ensures that the target global singleton row exists by adding it if it's
- * missing.
- *
- * @param {IDBTransaction} txn The version upgrade transaction for indexeddb
- */
-function ensureTargetGlobalExists(
-  txn: SimpleDbTransaction
-): PersistencePromise<DbTargetGlobal> {
-  const globalStore = txn.store<DbTargetGlobalKey, DbTargetGlobal>(
-    DbTargetGlobal.store
+  const metadata = new DbTargetGlobal(
+    /*highestTargetId=*/ 0,
+    /*lastListenSequenceNumber=*/ 0,
+    SnapshotVersion.MIN.toTimestamp(),
+    /*targetCount=*/ 0
   );
-  return globalStore.get(DbTargetGlobal.key).next(metadata => {
-    if (metadata != null) {
-      return PersistencePromise.resolve(metadata);
-    } else {
-      metadata = new DbTargetGlobal(
-        /*highestTargetId=*/ 0,
-        /*lastListenSequenceNumber=*/ 0,
-        SnapshotVersion.MIN.toTimestamp(),
-        /*targetCount=*/ 0
-      );
-      return globalStore.put(DbTargetGlobal.key, metadata).next(() => metadata);
-    }
-  });
+  return globalStore.put(DbTargetGlobal.key, metadata);
 }
 
 /**