+++ packages/firestore/CHANGELOG.md
@@ -1,4 +1,12 @@
 # Unreleased
+- [fixed] Fixed an issue where documents could be cached as deleted when
+  they should not have been, leading to queries returning fewer results than
+  they should (firebase/firebase-ios-sdk#1548).
+- [feature] Added `firebase.firestore.FieldValue.arrayUnion()` and
+  `firebase.firestore.FieldValue.arrayRemove()` to atomically add and remove
+  elements from an array field in a document.
+- [feature] Added `'array-contains'` query operator for use with `.where()` to
+  find documents where an array field contains a specific element.
 
 # 0.5.0
 - [changed] Merged the `includeQueryMetadataChanges` and